refactor(core): move dart-related method definitions to `structure.rs`
